chore: move package from org.dashevo to org.dash.platform
