--- src/main/kotlin/org/dashevo/dapiclient/model/JsonRPCRequest.kt
@@ -1,8 +1,0 @@
-package org.dashevo.dapiclient.model
-
-data class JsonRPCRequest<T: Any>(
-        val method: String,
-        val params: T? = null,
-        val jsonrpc: String = "2.0",
-        val id: String = "1"
-)
