--- src/main/kotlin/org/dashevo/dapiclient/model/JsonRPCResponse.kt
@@ -1,8 +1,0 @@
-package org.dashevo.dapiclient.model
-
-open class JsonRPCResponse<T: Any>(
-        val result: T? = null,
-        val id: String,
-        val jsonrpc: String,
-        val error: Map<String, Any>?
-)
