--- src/main/kotlin/org/dashevo/dapiclient/provider/DAPIMasternode.kt
@@ -1,9 +1,0 @@
-/**
- * Copyright (c) 2020-present, Dash Core Group
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-package org.dashevo.dapiclient.provider
-
-abstract class DAPIMasternode(val address: DAPIAddress)
